Adding modularity to the code
